Bugfix and improvements.  - Date formattings in search results improved.  - Alliance sings in White Footer section removed.  - Confrmation dialog added on research delete button click.  - Single quotes replaced in search results links which cause links to be broken.  - Removed commented codes.  - Resarch list ordered by name.

+++ src/com/mresearch/databank/client/presenters/AdminResearchDetailedPresenter.java
@@ -163,7 +163,9 @@ public class AdminResearchDetailedPresenter implements Presenter{
 			edit_display.getDeleteBtn().addClickHandler(new ClickHandler() {
 				@Override
 				public void onClick(ClickEvent arg0) {
-					deleteResearch(edit_display.getResearchID());
+          if (Window.confirm("Исследование будет безвозвратно удалено. Продолжить?")) {
+            deleteResearch(edit_display.getResearchID());
+          }
 				}
 			});
 	}
